+++ src/Common/EngineCore/CoreServiceCollectionExtensions.cs
@@ -141,6 +141,7 @@ public static class CoreServiceCollectionExtensions
         services.TryAddComponentCollection<QuantityComponentCollection>();
         services.TryAddComponentCollection<ItemUseComponentCollection>();
         services.TryAddComponentCollection<NpcFlagsComponentCollection>();
+        services.TryAddComponentCollection<PlayerFlagsComponentCollection>();
         services.TryAddComponentCollection<UseRangeComponentCollection>();
         services.TryAddComponentCollection<HealthComponentCollection>();
         services.TryAddComponentCollection<StaminaComponentCollection>();