+++ src/Common/EngineCore/Entities/AbstractEntityBuilder.cs
@@ -48,6 +48,7 @@ public abstract class AbstractEntityBuilder : IEntityBuilder
     private readonly NpcFlagsComponentCollection npcFlags;
     protected readonly KinematicsComponentCollection Kinematics;
     protected readonly bool load;
+    private readonly PlayerFlagsComponentCollection playerFlags;
     protected readonly NameComponentCollection names;
     protected readonly OrientationComponentCollection orientations;
     protected readonly ParentComponentCollection parents;
@@ -95,6 +96,7 @@ public abstract class AbstractEntityBuilder : IEntityBuilder
         QuantityComponentCollection quantities,
         ItemUseComponentCollection itemUses,
         NpcFlagsComponentCollection npcFlags,
+        PlayerFlagsComponentCollection playerFlags,
         UseRangeComponentCollection useRanges,
         HealthComponentCollection healths,
         StaminaComponentCollection staminas,
@@ -140,6 +142,7 @@ public abstract class AbstractEntityBuilder : IEntityBuilder
         this.levels = levels;
         this.experiences = experiences;
         this.radiantDatas = radiantDatas;
+        this.playerFlags = playerFlags;
 
         if (entityId is >= EntityConstants.FirstTemplateEntityId and <= EntityConstants.LastTemplateEntityId)
         {
@@ -598,6 +601,18 @@ public abstract class AbstractEntityBuilder : IEntityBuilder
     public IEntityBuilder WithoutNpcFlags()
     {
         npcFlags.RemoveComponent(entityId, load);
+        return this;
+    }
+
+    public IEntityBuilder PlayerFlags(PlayerFlag playerFlag)
+    {
+        playerFlags.AddOrUpdateComponent(entityId, playerFlag, load);
+        return this;
+    }
+
+    public IEntityBuilder WithoutPlayerFlags()
+    {
+        playerFlags.RemoveComponent(entityId, load);
         return this;
     }