+++ src/Server/Persistence/Database/Sqlite/Queries/SqliteRetrieveEntityQuery.cs
@@ -39,7 +39,7 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
                 manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
                 statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck,
                 equipmentType, level, experience,
-                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2)
+                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2, playerFlags)
 	        AS (
 		        SELECT id, template_id, x, y, z, frontTileId, topTileId, playerCharacter, name, account, parent, 
                         drawableX, drawableY, animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity,
@@ -50,7 +50,7 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
                         manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
                         statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck,
                         equipmentType, level, experience,
-                        radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2
+                        radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2, playerFlags
 		        FROM EntityWithComponents WHERE id = @Id
 	        UNION ALL
             	SELECT ec.id, ec.template_id, ec.x, ec.y, ec.z, ec.frontTileId, ec.topTileId, ec.playerCharacter,
@@ -64,7 +64,7 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
                         ec.manaValue, ec.manaMaxValue, ec.manaChangeRate, ec.manaChangeInterval,
                         ec.statsStrength, ec.statsDefense, ec.statsAgility, ec.statsIntelligence, ec.statsWisdom, ec.statsCharisma, ec.statsLuck,
                         ec.equipmentType, ec.level, ec.experience,
-                        ec.radiantCategory, ec.radiantFunction, ec.radiantParam0, ec.radiantParam1, ec.radiantParam2
+                        ec.radiantCategory, ec.radiantFunction, ec.radiantParam0, ec.radiantParam1, ec.radiantParam2, ec.playerFlags
 		        FROM EntityWithComponents ec, EntityTree et
         			WHERE ec.parent = et.id
 	        )
@@ -78,7 +78,7 @@ public sealed class SqliteRetrieveEntityQuery : IRetrieveEntityQuery
                 et.manaValue, et.manaMaxValue, et.manaChangeRate, et.manaChangeInterval,
                 et.statsStrength, et.statsDefense, et.statsAgility, et.statsIntelligence, et.statsWisdom, et.statsCharisma, et.statsLuck,
                 et.equipmentType, et.level, et.experience,
-                et.radiantCategory, et.radiantFunction, et.radiantParam0, et.radiantParam1, et.radiantParam2
+                et.radiantCategory, et.radiantFunction, et.radiantParam0, et.radiantParam1, et.radiantParam2, et.playerFlags
             FROM EntityTree et
             LEFT JOIN EntityKeyValue kv ON kv.entity_id = et.id
             ORDER BY et.parent NULLS LAST";