+++ src/Server/Persistence/Database/Sqlite/Queries/SqliteRetrieveRangeQuery.cs
@@ -41,7 +41,7 @@ public sealed class SqliteRetrieveRangeQuery : IRetrieveRangeQuery
                 manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
                 statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck,
                 equipmentType, level, experience,
-                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2)
+                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2, playerFlags)
 	        AS (
 	        	SELECT id, template_id, x, y, z, frontTileId, topTileId, playerCharacter, name, account, parent,
                         drawableX, drawableY, animatedSprite, orientation, admin, castBlockShadows, plsRadius, plsIntensity,
@@ -52,7 +52,7 @@ public sealed class SqliteRetrieveRangeQuery : IRetrieveRangeQuery
                         manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
                         statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck,
                         equipmentType, level, experience,
-                        radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2
+                        radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2, playerFlags
 	        		FROM EntityWithComponents
 	        		WHERE x >= @X1 AND x < @X2
 	        		  AND y >= @Y1 AND y < @Y2
@@ -70,7 +70,7 @@ public sealed class SqliteRetrieveRangeQuery : IRetrieveRangeQuery
                         ec.manaValue, ec.manaMaxValue, ec.manaChangeRate, ec.manaChangeInterval,
                         ec.statsStrength, ec.statsDefense, ec.statsAgility, ec.statsIntelligence, ec.statsWisdom, ec.statsCharisma, ec.statsLuck,
                         ec.equipmentType, ec.level, ec.experience,
-                        ec.radiantCategory, ec.radiantFunction, ec.radiantParam0, ec.radiantParam1, ec.radiantParam2
+                        ec.radiantCategory, ec.radiantFunction, ec.radiantParam0, ec.radiantParam1, ec.radiantParam2, ec.playerFlags
 	        		FROM EntityWithComponents ec, EntityTree et
 	        		WHERE ec.parent = et.id 
                       AND ec.playerCharacter IS NULL
@@ -84,7 +84,7 @@ public sealed class SqliteRetrieveRangeQuery : IRetrieveRangeQuery
                 manaValue, manaMaxValue, manaChangeRate, manaChangeInterval,
                 statsStrength, statsDefense, statsAgility, statsIntelligence, statsWisdom, statsCharisma, statsLuck,
                 equipmentType, level, experience,
-                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2
+                radiantCategory, radiantFunction, radiantParam0, radiantParam1, radiantParam2, playerFlags
             FROM EntityTree 
             LEFT JOIN EntityKeyValue kv ON kv.entity_id = id
             ORDER BY parent NULLS LAST";